requirements.txt
# requirements.txt
mcp
requests
websocket-client
Pillow
python-dotenv